UI finished

+++ comp421project/src/main/java/test/MakeAndCancelAppointment.java
@@ -1,57 +1,72 @@
 package test;
 
 import java.sql.*;
-import java.sql.Time;
 import java.lang.System;
 import java.util.Date;
-import java.sql.SQLException;
+import java.util.Scanner;
 
 public class MakeAndCancelAppointment {
 
-    public static void MakeCustomAppointment(Statement statement, String starttime, String cemail, String endtime,
-            String appdate, String creditcardnumber, String nartistemail, String locationname) {
-        // make appointment with some custom starttime, endtime and apppointment date
-        if (nartistemail == "" || nartistemail == null) {
-            System.out.println("Nail Artist Email cannot be null or empty");
-            return;
-        }
-        if (cemail == "" || cemail == null) {
-            System.out.println("Customer email cannot be null or empty");
-            return;
-        }
-        if (locationname == "" || locationname == null) {
-            System.out.println("Location name cannot be null or empty");
-            return;
-        }
-        if (starttime == "" || starttime == null) {
-            System.out.println("Start Time cannot be null or empty");
-            return;
-        }
-        if (endtime == "" || endtime == null) {
-            System.out.println("End time cannot be null or empty");
-            return;
-        }
-        if (appdate == "" || appdate == null) {
-            System.out.println("Appointment Date cannot be null or empty");
-            return;
-        }
+  public static void MakeCustomAppointment(Scanner scanner, Statement statement) {
+    System.out.println("Enter the start time of the appointment (HH:MM:SS):");
+    Scanner scanner1 = new Scanner(System.in);
+    if (scanner1.hasNext()){
+      String starttime = scanner1.nextLine(); 
+      System.out.println("Enter the customer's email address:");
+      String cemail = scanner1.nextLine(); 
+      System.out.println("Enter the end time of the appointment (HH:MM:SS):");
+      String endtime = scanner1.nextLine(); 
+      System.out.println("Enter the date of the appointment (M/D/Y):");
+      String appdate = scanner1.nextLine(); 
+      System.out.println("Enter the credit card number of the customer:");
+      String creditcardnumber = scanner1.nextLine();
+      System.out.println("Enter the nail artist's email address:");
+      String nartistemail = scanner1.nextLine(); 
+      System.out.println("Enter the location of the appointment:");
+      String locationname = scanner1.nextLine();
+  // make appointment with some custom starttime, endtime and apppointment date
+  if (nartistemail == "" || nartistemail == null) {
+      System.out.println("Nail Artist Email cannot be null or empty");
+      return;
+  }
+  if (cemail == "" || cemail == null) {
+      System.out.println("Customer email cannot be null or empty");
+      return;
+  }
+  if (locationname == "" || locationname == null) {
+      System.out.println("Location name cannot be null or empty");
+      return;
+  }
+  if (starttime == "" || starttime == null) {
+      System.out.println("Start Time cannot be null or empty");
+      return;
+  }
+  if (endtime == "" || endtime == null) {
+      System.out.println("End time cannot be null or empty");
+      return;
+  }
+  if (appdate == "" || appdate == null) {
+      System.out.println("Appointment Date cannot be null or empty");
+      return;
+  }
 
-        int sqlCode = 0; // Variable to hold SQLCODE
-        String sqlState = "00000"; // Variable to hold SQLSTATE
-        try {
-            String insertSQLToAppointment = "INSERT INTO appointment VALUES ( \'" + starttime + "\',  \'" + cemail
-                    + "\',  \'" + endtime + "\',  \'" + appdate + "\',  \'" + creditcardnumber + "\', \'" + nartistemail
-                    + "\', \'" + locationname + "\') ";
-            System.out.println(insertSQLToAppointment);
-            statement.executeUpdate(insertSQLToAppointment);
-            System.out.println("DONE");
+  int sqlCode = 0; // Variable to hold SQLCODE
+  String sqlState = "00000"; // Variable to hold SQLSTATE
+  try {
+      String insertSQLToAppointment = "INSERT INTO appointment VALUES ( \'" + starttime + "\',  \'" + cemail
+              + "\',  \'" + endtime + "\',  \'" + appdate + "\',  \'" + creditcardnumber + "\', \'" + nartistemail
+              + "\', \'" + locationname + "\') ";
+      System.out.println(insertSQLToAppointment);
+      statement.executeUpdate(insertSQLToAppointment);
+      System.out.println("DONE");
 
-        } catch (SQLException e) {
-            sqlCode = e.getErrorCode(); // Get SQLCODE
-            sqlState = e.getSQLState(); // Get SQLSTATE
-            System.out.println("Code: " + sqlCode + "  sqlState: " + sqlState);
-        }
+  } catch (SQLException e) {
+      sqlCode = e.getErrorCode(); // Get SQLCODE
+      sqlState = e.getSQLState(); // Get SQLSTATE
+      System.out.println("Code: " + sqlCode + "  sqlState: " + sqlState);
     }
+   }
+ }
 
     public static void MakeAppointmentToday(Statement statement, String cemail, String creditcardnumber,
             String nartistemail, String locationname) {
@@ -95,33 +110,42 @@ public class MakeAndCancelAppointment {
         }
     }
 
-    public static void CancelAppointment(Statement statement, String starttime, String cemail, String appdate) {
-        int sqlCode = 0; // Variable to hold SQLCODE
-        String sqlState = "00000"; // Variable to hold SQLSTATE
-        if (starttime == "" || starttime == null) {
-            System.out.println("Start Time cannot be null or empty");
-            return;
-        }
-        if (cemail == "" || cemail == null) {
-            System.out.println("Customer email cannot be null or empty");
-            return;
-        }
-        if (appdate == "" || appdate == null) {
-            System.out.println("Appointment Date cannot be null or empty");
-            return;
-        }
-        try {
-            String deleteAppointment = "DELETE FROM appointment WHERE starttime = \'" + starttime + "\' AND cemail = \'"
-                    + cemail + "\' AND appdate = \'" + appdate + "\'";
-            System.out.println(deleteAppointment);
-            statement.executeUpdate(deleteAppointment);
-            System.out.println("DONE");
-
-        } catch (SQLException e) {
-            sqlCode = e.getErrorCode(); // Get SQLCODE
-            sqlState = e.getSQLState(); // Get SQLSTATE
-            System.out.println("Code: " + sqlCode + "  sqlState: " + sqlState);
-        }
-    }
+    public static void CancelAppointment(Scanner scanner, Statement statement) {
+      System.out.println("Enter the start time of the appointment (HH:MM:SS):");
+       Scanner scanner1 = new Scanner(System.in);
+       if (scanner1.hasNext()){
+          String starttime = scanner1.nextLine(); 
+          System.out.println("Enter the customer's email address:");
+          String cemail = scanner1.nextLine(); 
+          System.out.println("Enter the date of the appointment (M/D/Y):");
+          String appdate = scanner1.nextLine(); 
+          int sqlCode = 0; // Variable to hold SQLCODE
+          String sqlState = "00000"; // Variable to hold SQLSTATE
+          if (starttime == "" || starttime == null) {
+              System.out.println("Start Time cannot be null or empty");
+              return;
+          }
+          if (cemail == "" || cemail == null) {
+              System.out.println("Customer email cannot be null or empty");
+              return;
+          }
+          if (appdate == "" || appdate == null) {
+              System.out.println("Appointment Date cannot be null or empty");
+              return;
+          }
+          try {
+              String deleteAppointment = "DELETE FROM appointment WHERE starttime = \'" + starttime + "\' AND cemail = \'"
+                      + cemail + "\' AND appdate = \'" + appdate + "\'";
+              System.out.println(deleteAppointment);
+              statement.executeUpdate(deleteAppointment);
+              System.out.println("DONE");
+   
+          } catch (SQLException e) {
+              sqlCode = e.getErrorCode(); // Get SQLCODE
+              sqlState = e.getSQLState(); // Get SQLSTATE
+              System.out.println("Code: " + sqlCode + "  sqlState: " + sqlState);
+          }
+      }
+   }
 
 }
